--- app/src/main/java/co/touchlab/kampkit/android/design/ButtonStyles.kt
@@ -1,2 +1,0 @@
-package co.touchlab.kampkit.android.design
-
