+++ app/src/main/java/co/touchlab/kampkit/android/design/styles/ButtonStyles.kt
@@ -1,0 +1,5 @@
+package co.touchlab.kampkit.android.design.styles
+
+object ButtonStyles{
+    
+}
