+++ app/src/main/java/co/touchlab/kampkit/android/design/styles/TextStyles.kt
@@ -1,9 +1,10 @@
-package co.touchlab.kampkit.android.design
+package co.touchlab.kampkit.android.design.styles
 
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
+import co.touchlab.kampkit.android.design.TextSize
 
 object TextStyles {
     val xxl = TextStyle(
